requirements.txt
aiodns==3.0.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
dataclasses-json==0.5.14
exceptiongroup==1.1.3
fastapi==0.103.1
Flask==2.2.5
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.33
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==2.0.2
grpcio==1.57.0
grpcio-tools==1.57.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.251
langsmith==0.0.32
loguru==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
omegaconf==2.3.0
openai==0.28.1
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
portalocker==2.7.0
protobuf==4.24.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pypdf==3.15.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-liquid==1.9.4
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
qdrant-client==1.5.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rpds-py==0.10.0
rsa==4.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
streamlit==1.26.0
swagger-ui-bundle==0.0.9
tenacity==8.2.3
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
uvicorn==0.23.2
validators==0.21.2
watchdog==3.0.0
Werkzeug==2.2.3
win32-setctime==1.1.0
yarl==1.9.2
zipp==3.16.2
